small fixes and added import-export module

+++ FinanceServer/simplifiedFinance/views/accountViews.py
@@ -20,33 +20,37 @@ def getAccounts(request):
 @require_http_methods(["POST"])
 @csrf_exempt
 def createAccount(request):
-    if request.user.is_authenticated:
-        if request.user.is_superuser:
+    try:
+        if request.user.is_authenticated and request.user.is_superuser:
             name = request.POST.get("accountName")
             accountNumber = request.POST.get("accountNumber")
             balance = request.POST.get("accountBalance")
-            account = SubAccount(name=name, accountNumber=accountNumber, balance=balance)
-            account.save()
-            return JsonResponse({"status": "success", "id": account.id})
+
+            if SubAccount.objects.filter(name=name).exists():
+                raise IntegrityError("Account already exists")
+            else:
+                account = SubAccount(name=name, accountNumber=accountNumber, balance=balance)
+                account.save()
+
+                return JsonResponse({"status": "success", "id": account.id})
         else:
             return JsonResponse({"status": "error", "message": "User is not a superuser"})
-    else:
-        return JsonResponse({"status": "error", "message": "User is not authenticated"})
+    except IntegrityError:
+        return JsonResponse({"status": "error", "message": "Account already exists"})
+    except Exception as e:
+        return JsonResponse({"status": "error", "message": str(e)})
 
 
 @require_http_methods(["DELETE"])
 @csrf_exempt
 def deleteAccount(request, name):
     try:
-        if request.user.is_authenticated:
-            if request.user.is_superuser:
-                account = SubAccount.objects.get(name=name)
-                account.delete()
-                return JsonResponse({"status": "success"})
-            else:
-                return JsonResponse({"status": "error", "message": "User is not a superuser"})
+        if request.user.is_authenticated and request.user.is_superuser:
+            account = SubAccount.objects.get(name=name)
+            account.delete()
+            return JsonResponse({"status": "success"})
         else:
-            return JsonResponse({"status": "error", "message": "User is not authenticated"})
+            return JsonResponse({"status": "error", "message": "User is not a superuser"})
     except SubAccount.DoesNotExist:
         return JsonResponse({"status": "error", "message": "Account does not exist"})
     except IntegrityError:
