+++ FinanceServer/simplifiedFinance/views/projectViews.py
@@ -20,35 +20,40 @@ def getProjects(request):
 @require_http_methods(["POST"])
 @csrf_exempt
 def createProject(request):
-    if request.user.is_authenticated:
-        if request.user.is_superuser:
+    try:
+        if request.user.is_authenticated and request.user.is_superuser:
             name = request.POST.get("projectName")
             description = request.POST.get("projectDescription")
             startDate = request.POST.get("projectStartDate")
             endDate = request.POST.get("projectEndDate")
             status = request.POST.get("projectStatus")
-            project = Project(name=name, description=description, startDate=startDate, endDate=endDate, status=status)
-            project.save()
+
+            if Project.objects.filter(name=name).exists():
+                raise IntegrityError("Project already exists")
+            else:
+                project = Project(name=name, description=description, startDate=startDate, endDate=endDate, status=status)
+                project.save()
+
             return JsonResponse({"status": "success", "id": project.id})
         else:
             return JsonResponse({"status": "error", "message": "User is not a superuser"})
-    else:
-        return JsonResponse({"status": "error", "message": "User is not authenticated"})
+    except IntegrityError:
+        return JsonResponse({"status": "error", "message": "Project already exists"})
+    except Exception as e:
+        return JsonResponse({"status": "error", "message": str(e)})
 
 
 @require_http_methods(["DELETE"])
 @csrf_exempt
 def deleteProject(request, name):
     try:
-        if request.user.is_authenticated:
-            if request.user.is_superuser:
-                project = Project.objects.get(name=name)
-                project.delete()
-                return JsonResponse({"status": "success"})
-            else:
-                return JsonResponse({"status": "error", "message": "User is not a superuser"})
+        if request.user.is_authenticated and request.user.is_superuser:
+
+            project = Project.objects.get(name=name)
+            project.delete()
+            return JsonResponse({"status": "success"})
         else:
-            return JsonResponse({"status": "error", "message": "User is not authenticated"})
+            return JsonResponse({"status": "error", "message": "User is not a superuser"})
     except Project.DoesNotExist:
         return JsonResponse({"status": "error", "message": "Project does not exist"})
     except IntegrityError:
@@ -61,19 +66,14 @@ def deleteProject(request, name):
 @csrf_exempt
 def endProject(request, name):
     try:
-        if request.user.is_authenticated:
-            if request.user.is_superuser:
-                project = Project.objects.get(name=name)
-                project.status = "Completed"
-                project.save()
-                return JsonResponse({"status": "success"})
-            else:
-                return JsonResponse({"status": "error", "message": "User is not a superuser"})
+        if request.user.is_authenticated and request.user.is_superuser:
+            project = Project.objects.get(name=name)
+            project.status = "Completed"
+            project.save()
+            return JsonResponse({"status": "success"})
         else:
-            return JsonResponse({"status": "error", "message": "User is not authenticated"})
+            return JsonResponse({"status": "error", "message": "User is not a superuser"})
     except Project.DoesNotExist:
         return JsonResponse({"status": "error", "message": "Project does not exist"})
-    except IntegrityError:
-        return JsonResponse({"status": "error", "message": "Project is in use"})
     except Exception as e:
         return JsonResponse({"status": "error", "message": str(e)})
